support nodejs esm (#68)
* support nodejs esm env

* fix ssr in node esm env

* chore: publish alpha

* WIP: tests

* fix import client and clientTypes

* update lock

* WIP: tests

* tests cleanup

* fix test: kill ESRCH

* improve tests

* improve test setup

* add comment

* cleanup

* improve test timeout setup

+++ test-setup/utils/index.ts
@@ -11,6 +11,8 @@ export * from '@playwright/test'
 
 export type TestOptions = {
   vitePagesMode: 'serve' | 'build' | 'ssr'
+  skipPrepare: boolean
+  beforeStartViteServer: (...args: any[]) => any
 }
 
 export const test = base.extend<
@@ -21,13 +23,25 @@ export const test = base.extend<
       path: string
       restore: (subPath?: string) => Promise<void>
     }
-    server: { port: number }
     fsUtils: ReturnType<typeof getFsUtils>
+    server: { port: number }
   } & TestOptions
 >({
   vitePagesMode: ['serve', { option: true, scope: 'worker' }],
+  skipPrepare: [false, { option: true, scope: 'worker' }],
+  beforeStartViteServer: [
+    async ({}, use) => {
+      // beforeStartViteServer default to be no-op
+      await use(() => {})
+    },
+    { option: true, scope: 'worker' },
+  ],
   testPlayground: [
-    async ({}, use, workerInfo) => {
+    async ({ skipPrepare }, use, workerInfo) => {
+      if (skipPrepare) {
+        await use(null as any)
+        return
+      }
       const testDir = workerInfo.project.testDir
       if (!(await fs.pathExists(testDir)))
         throw new Error(`testDir not exists: ${testDir}`)
@@ -44,8 +58,27 @@ export const test = base.extend<
     },
     { scope: 'worker', auto: true },
   ],
+  fsUtils: [
+    async ({ skipPrepare, testPlayground }, use) => {
+      if (skipPrepare) {
+        await use(null as any)
+        return
+      }
+      await use(getFsUtils(testPlayground.path))
+    },
+    { scope: 'worker' },
+  ],
   server: [
-    async ({ testPlayground, vitePagesMode }, use, workerInfo) => {
+    async (
+      { testPlayground, vitePagesMode, beforeStartViteServer, skipPrepare },
+      use,
+      workerInfo
+    ) => {
+      if (skipPrepare) {
+        await use(null as any)
+        return
+      }
+      await beforeStartViteServer?.()
       const vars: { port?: number; subprocess?: any } = {}
       try {
         if (vitePagesMode === 'serve') {
@@ -61,30 +94,33 @@ export const test = base.extend<
         if (vars.subprocess) killProcess(vars.subprocess)
       }
     },
-    { scope: 'worker', auto: true },
-  ],
-  baseURL: async ({ server }, use) => {
-    await use(`http://localhost:${server.port}`)
-  },
-  page: async ({ baseURL, page, server }, use) => {
-    // double check if this fixture works
-    if (baseURL !== `http://localhost:${server.port}`)
-      throw new Error('unexpected baseURL')
-
-    // const res = await axios.get(baseURL, {
-    //   timeout: 5000,
-    //   headers: { Accept: 'text/html' },
-    // })
-    // console.log('@@baseURL.data', res.data)
-
-    await page.goto(baseURL)
-    await use(page)
-  },
-  fsUtils: [
-    async ({ testPlayground }, use, workerInfo) => {
-      await use(getFsUtils(testPlayground.path))
+    {
+      scope: 'worker',
+      auto: true,
+      timeout: 60 * 1000,
     },
-    { scope: 'worker' },
+  ],
+  baseURL: async ({ skipPrepare, server, baseURL }, use) => {
+    await use(skipPrepare ? baseURL : `http://localhost:${server.port}`)
+  },
+  // if you are running tests in wsl, and you get error:
+  // Timeout exceeded while running fixture "page" setup
+  // You may have forgotten to start VcXsrv (XLaunch)
+  // Ref: how to run e2e tests in wsl:
+  // https://shouv.medium.com/how-to-run-cypress-on-wsl2-989b83795fb6
+  page: [
+    async ({ baseURL, page }, use) => {
+      if (baseURL) {
+        // const res = await axios.get(baseURL, {
+        //   timeout: 5000,
+        //   headers: { Accept: 'text/html' },
+        // })
+        // console.log('@@baseURL.data', res.data)
+        await page.goto(baseURL)
+      }
+      await use(page)
+    },
+    { scope: 'test', timeout: 60 * 1000 },
   ],
 })
 
@@ -96,8 +132,10 @@ export async function killProcess(subprocess: ExecaChildProcess) {
     // ref: https://github.com/vitejs/vite/blob/f9b5c14c42bf0a5c7d4ca4b53160047306fb07c5/playground/test-utils.ts#L281
     commandSync(`taskkill /pid ${subprocess.pid} /T /F`)
   } else if (subprocess.pid) {
-    // https://stackoverflow.com/a/49842576
-    process.kill(-subprocess.pid)
+    if (subprocess.exitCode === null) {
+      // https://stackoverflow.com/a/49842576
+      process.kill(-subprocess.pid)
+    }
   } else {
     subprocess.kill()
   }
